Fri, Apr 8, 2022, 3:56 PM +00:00

+++ readability/readability.c
@@ -91,9 +91,9 @@ int count_sentences(string text)
 
 int grade_calculator(letters, words, sentences)
 {
-    float L = (float)letters / (float)words * 100;
-    float S = (float)sentences / (float)words * 100;
-    float index = 0.0588 * L - 0.296 * S - 15.8;
+    double L = (float)letters / (float)words * 100;
+    double S = (float)sentences / (float)words * 100;
+    double index = 0.0588 * L - 0.296 * S - 15.8;
     printf("index before round: %f", index);
     float round(float index);
     printf("index after round: %f", index);
